+++ lib/core/constants/end-points.dart
@@ -1,4 +1,4 @@
 class EndPoints {
   static const baseUrl = 'https://api.github.com';
-  static const githubProfile = '/users/uzairleo';
+  static const githubProfile = '/users';
 }
