+++ lib/core/models/exception/response-exception.dart
@@ -1,0 +1,5 @@
+class ResponseException implements Exception {
+  String cause;
+
+  ResponseException(this.cause);
+}
