+++ lib/core/models/github-user.dart
@@ -16,13 +16,13 @@ class GithubUser {
   String eventsUrl;
   String receivedEventsUrl;
   String type;
-  bool siteAdmin;
+  // bool siteAdmin;
   String name;
   String company;
   String blog;
   String location;
-  Null email;
-  Null hireable;
+  // Null email;
+  // Null hireable;
   String bio;
   String twitterUsername;
   int publicRepos;
@@ -50,13 +50,10 @@ class GithubUser {
       this.eventsUrl,
       this.receivedEventsUrl,
       this.type,
-      this.siteAdmin,
       this.name,
       this.company,
       this.blog,
       this.location,
-      this.email,
-      this.hireable,
       this.bio,
       this.twitterUsername,
       this.publicRepos,
@@ -84,13 +81,10 @@ class GithubUser {
     eventsUrl = json['events_url'];
     receivedEventsUrl = json['received_events_url'];
     type = json['type'];
-    siteAdmin = json['site_admin'];
     name = json['name'];
     company = json['company'];
     blog = json['blog'];
     location = json['location'];
-    email = json['email'];
-    hireable = json['hireable'];
     bio = json['bio'];
     twitterUsername = json['twitter_username'];
     publicRepos = json['public_repos'];
@@ -120,13 +114,10 @@ class GithubUser {
     data['events_url'] = this.eventsUrl;
     data['received_events_url'] = this.receivedEventsUrl;
     data['type'] = this.type;
-    data['site_admin'] = this.siteAdmin;
     data['name'] = this.name;
     data['company'] = this.company;
     data['blog'] = this.blog;
     data['location'] = this.location;
-    data['email'] = this.email;
-    data['hireable'] = this.hireable;
     data['bio'] = this.bio;
     data['twitter_username'] = this.twitterUsername;
     data['public_repos'] = this.publicRepos;
