+++ pubspec.yaml
@@ -21,20 +21,17 @@ environment:
   sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
-  
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
+  dio: ^4.0.0
+  dio_http_cache:
+    git:
+      url: https://github.com/hurshi/dio-http-cache
   flutter:
     sdk: flutter
   font_awesome_flutter: ^9.1.0
   page_transition: ^1.1.4
-  dio: ^4.0.0
   provider: ^5.0.0
-  dio_http_cache: 
-    git:
-      url: https://github.com/hurshi/dio-http-cache 
-   
+
   simple_animations: ^1.1.3
 
   url_launcher: null
@@ -43,6 +40,7 @@ dev_dependencies:
   flip_card: ^0.5.0
   flutter_test:
     sdk: flutter
+  get: ^4.3.8
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
